% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agnostic.aov.R
\name{agnostic.aov}
\alias{agnostic.aov}
\title{Agnostic Analysis of Variance Model}
\usage{
agnostic.aov(formula, data = NULL, projections = FALSE, qr = TRUE,
  contrasts = NULL, ...)
}
\arguments{
\item{formula}{A formula specifying the model.}

\item{data}{A data frame in which the variables specified in the formula will be found. If missing, the variables are searched for in the standard way.}

\item{projections}{Logical flag: should the projections be returned?}

\item{qr}{Logical flag: should the QR decomposition be returned?}

\item{contrasts}{A list of contrasts to be used for some of the factors in the formula. These are not used for any Error term, and supplying contrasts for factors only in the Error term will give a warning.}

\item{...}{Arguments to be passed to lm, such as subset or na.action. See 'Details' about weights.}
}
\value{
An object of class c("agnostic.aov", "lm") or for multiple responses of class c("maov", "agnostic.aov", "mlm", "lm") or for multiple error strata of class c("aovlist", "listof"). There are print and summary methods available for these.
}
\description{
Fit an analysis of variance model under agnostic perspective.
}
\examples{
#Test data
obs <- c(54, 60, 59, 45, 40, 55, 47, 33, 51, 66, 44, 34, 36, 61, 49, 48, 50, 54, 62, 42, 48, 61, 60, 44)
obs2 <- rnorm(24)
groups <- as.factor(rep(c("A", "B", "c", "D"), 6))
groups2 <- as.factor(rep(c("A", "D", "B", "C"), 6))

mod1 <- agnostic.aov(obs ~ groups)
summary(mod1)
mod2 <- agnostic.aov(cbind(obs, obs2) ~ groups * groups2)
summary(mod2)
}
