% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_home_dir.R
\name{get_home_dir}
\alias{get_home_dir}
\title{GET_HOME_DIR}
\usage{
get_home_dir()
}
\value{
the home directory string of your file system
}
\description{
This funtion returns the home directory of your file system regardless of what operating system you're using.
}
\examples{
get_home_dir()
}
\keyword{file}
