% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.agnostic.aov.R
\name{summary.agnostic.aov}
\alias{summary.agnostic.aov}
\title{Summarize an Agnostic Analysis of Variance Model}
\usage{
\method{summary}{agnostic.aov}(object, intercept = FALSE, split,
  expand.split = TRUE, keep.zero.df = TRUE, ...)
}
\arguments{
\item{object}{object of class "agnostic.aov".}

\item{intercept}{logical: should intercept terms be included?}

\item{split}{an optional named list, with names corresponding to terms in the model. Each component is itself a list with integer components giving contrasts whose contributions are to be summed.}

\item{expand.split}{logical: should the split apply also to interactions involving the factor?}

\item{keep.zero.df}{logical: should terms with no degrees of freedom be included?}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
An object of class c("summary.aov", "listof") or "summary.aovlist" respectively.

For fits with a single stratum the result will be a list of agnostic ANOVA tables, one for each response (even if there is only one response): the tables are of class "anova" inheriting from class "data.frame". They have columns "Df", "Sum Sq", "Mean Sq", as well as "F value", "P.H0" and "P.H1" if there are non-zero residual degrees of freedom. There is a row for each term in the model, plus one for "Residuals" if there are any.

For multistratum fits the return value is a list of such summaries, one for each stratum.
}
\description{
Summarize an Agnostic Analysis of Variance Model
}
\examples{
#Test data
obs <- c(54, 60, 59, 45, 40, 55, 47, 33, 51, 66, 44, 34, 36, 61,
49, 48, 50, 54, 62, 42, 48, 61, 60, 44)
obs2 <- rnorm(24)
groups <- as.factor(rep(c("A", "B", "c", "D"), 6))
groups2 <- as.factor(rep(c("A", "D", "B", "C"), 6))

mod1 <- agnostic.aov(obs ~ groups)
summary(mod1)
mod2 <- agnostic.aov(cbind(obs, obs2) ~ groups * groups2)
summary(mod2)
}
